+++ Nakama/ApiClient.gen.cs
@@ -325,7 +325,7 @@ namespace Nakama
         IEnumerable<IApiAccountDevice> Devices { get; }
 
         /// <summary>
-        /// The UNIX time when the user's account was disabled/banned.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the user's account was disabled/banned.
         /// </summary>
         string DisableTime { get; }
 
@@ -340,7 +340,7 @@ namespace Nakama
         IApiUser User { get; }
 
         /// <summary>
-        /// The UNIX time when the user's email was verified.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the user's email was verified.
         /// </summary>
         string VerifyTime { get; }
 
@@ -550,7 +550,7 @@ namespace Nakama
         string Email { get; }
 
         /// <summary>
-        /// A password for the user account.
+        /// A password for the user account.  Ignored with unlink operations.
         /// </summary>
         string Password { get; }
 
@@ -687,7 +687,7 @@ namespace Nakama
     }
 
     /// <summary>
-    /// Send Apple's Game Center account credentials to the server. Used with authenticate/link/unlink.
+    /// Send Apple's Game Center account credentials to the server. Used with authenticate/link/unlink.  https://developer.apple.com/documentation/gamekit/gklocalplayer/1515407-generateidentityverificationsign
     /// </summary>
     public interface IApiAccountGameCenter
     {
@@ -896,7 +896,7 @@ namespace Nakama
         string Content { get; }
 
         /// <summary>
-        /// The UNIX time when the message was created.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the message was created.
         /// </summary>
         string CreateTime { get; }
 
@@ -926,7 +926,7 @@ namespace Nakama
         string SenderId { get; }
 
         /// <summary>
-        /// The UNIX time when the message was last updated.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the message was last updated.
         /// </summary>
         string UpdateTime { get; }
 
@@ -1311,7 +1311,7 @@ namespace Nakama
     {
 
         /// <summary>
-        /// The friend status.
+        /// The friend status.  one of "Friend.State".
         /// </summary>
         int State { get; }
 
@@ -1406,7 +1406,7 @@ namespace Nakama
         string AvatarUrl { get; }
 
         /// <summary>
-        /// The UNIX time when the group was created.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the group was created.
         /// </summary>
         string CreateTime { get; }
 
@@ -1456,7 +1456,7 @@ namespace Nakama
         bool Open { get; }
 
         /// <summary>
-        /// The UNIX time when the group was last updated.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the group was last updated.
         /// </summary>
         string UpdateTime { get; }
     }
@@ -1619,12 +1619,12 @@ namespace Nakama
     {
 
         /// <summary>
-        /// The UNIX time when the leaderboard record was created.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the leaderboard record was created.
         /// </summary>
         string CreateTime { get; }
 
         /// <summary>
-        /// The UNIX time when the leaderboard record expires.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the leaderboard record expires.
         /// </summary>
         string ExpiryTime { get; }
 
@@ -1669,7 +1669,7 @@ namespace Nakama
         string Subscore { get; }
 
         /// <summary>
-        /// The UNIX time when the leaderboard record was updated.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the leaderboard record was updated.
         /// </summary>
         string UpdateTime { get; }
 
@@ -1772,6 +1772,11 @@ namespace Nakama
         string PrevCursor { get; }
 
         /// <summary>
+        /// The total number of ranks available.
+        /// </summary>
+        string RankCount { get; }
+
+        /// <summary>
         /// A list of leaderboard records.
         /// </summary>
         IEnumerable<IApiLeaderboardRecord> Records { get; }
@@ -1796,6 +1801,10 @@ namespace Nakama
         public string PrevCursor { get; set; }
 
         /// <inheritdoc />
+        [DataMember(Name="rank_count"), Preserve]
+        public string RankCount { get; set; }
+
+        /// <inheritdoc />
         [IgnoreDataMember]
         public IEnumerable<IApiLeaderboardRecord> Records => _records ?? new List<ApiLeaderboardRecord>(0);
         [DataMember(Name="records"), Preserve]
@@ -1807,6 +1816,7 @@ namespace Nakama
             output = string.Concat(output, "NextCursor: ", NextCursor, ", ");
             output = string.Concat(output, "OwnerRecords: [", string.Join(", ", OwnerRecords), "], ");
             output = string.Concat(output, "PrevCursor: ", PrevCursor, ", ");
+            output = string.Concat(output, "RankCount: ", RankCount, ", ");
             output = string.Concat(output, "Records: [", string.Join(", ", Records), "], ");
             return output;
         }
@@ -2015,7 +2025,7 @@ namespace Nakama
         string Content { get; }
 
         /// <summary>
-        /// The UNIX time when the notification was created.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the notification was created.
         /// </summary>
         string CreateTime { get; }
 
@@ -2423,7 +2433,7 @@ namespace Nakama
         string Collection { get; }
 
         /// <summary>
-        /// The UNIX time when the object was created.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the object was created.
         /// </summary>
         string CreateTime { get; }
 
@@ -2443,7 +2453,7 @@ namespace Nakama
         int PermissionWrite { get; }
 
         /// <summary>
-        /// The UNIX time when the object was last updated.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the object was last updated.
         /// </summary>
         string UpdateTime { get; }
 
@@ -2531,9 +2541,19 @@ namespace Nakama
         string Collection { get; }
 
         /// <summary>
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the object was created.
+        /// </summary>
+        string CreateTime { get; }
+
+        /// <summary>
         /// The key of the object within the collection.
         /// </summary>
         string Key { get; }
+
+        /// <summary>
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the object was last updated.
+        /// </summary>
+        string UpdateTime { get; }
 
         /// <summary>
         /// The owner of the object.
@@ -2555,8 +2575,16 @@ namespace Nakama
         public string Collection { get; set; }
 
         /// <inheritdoc />
+        [DataMember(Name="create_time"), Preserve]
+        public string CreateTime { get; set; }
+
+        /// <inheritdoc />
         [DataMember(Name="key"), Preserve]
         public string Key { get; set; }
+
+        /// <inheritdoc />
+        [DataMember(Name="update_time"), Preserve]
+        public string UpdateTime { get; set; }
 
         /// <inheritdoc />
         [DataMember(Name="user_id"), Preserve]
@@ -2570,7 +2598,9 @@ namespace Nakama
         {
             var output = "";
             output = string.Concat(output, "Collection: ", Collection, ", ");
+            output = string.Concat(output, "CreateTime: ", CreateTime, ", ");
             output = string.Concat(output, "Key: ", Key, ", ");
+            output = string.Concat(output, "UpdateTime: ", UpdateTime, ", ");
             output = string.Concat(output, "UserId: ", UserId, ", ");
             output = string.Concat(output, "Version: ", Version, ", ");
             return output;
@@ -2787,7 +2817,7 @@ namespace Nakama
         int Category { get; }
 
         /// <summary>
-        /// The UNIX time when the tournament was created.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the tournament was created.
         /// </summary>
         string CreateTime { get; }
 
@@ -2807,7 +2837,7 @@ namespace Nakama
         int EndActive { get; }
 
         /// <summary>
-        /// The UNIX time when the tournament will be stopped.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the tournament will be stopped.
         /// </summary>
         string EndTime { get; }
 
@@ -2862,7 +2892,7 @@ namespace Nakama
         int StartActive { get; }
 
         /// <summary>
-        /// The UNIX time when the tournament will start.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the tournament will start.
         /// </summary>
         string StartTime { get; }
 
@@ -3047,6 +3077,11 @@ namespace Nakama
         string PrevCursor { get; }
 
         /// <summary>
+        /// The total number of ranks available.
+        /// </summary>
+        string RankCount { get; }
+
+        /// <summary>
         /// A list of tournament records.
         /// </summary>
         IEnumerable<IApiLeaderboardRecord> Records { get; }
@@ -3071,6 +3106,10 @@ namespace Nakama
         public string PrevCursor { get; set; }
 
         /// <inheritdoc />
+        [DataMember(Name="rank_count"), Preserve]
+        public string RankCount { get; set; }
+
+        /// <inheritdoc />
         [IgnoreDataMember]
         public IEnumerable<IApiLeaderboardRecord> Records => _records ?? new List<ApiLeaderboardRecord>(0);
         [DataMember(Name="records"), Preserve]
@@ -3082,6 +3121,7 @@ namespace Nakama
             output = string.Concat(output, "NextCursor: ", NextCursor, ", ");
             output = string.Concat(output, "OwnerRecords: [", string.Join(", ", OwnerRecords), "], ");
             output = string.Concat(output, "PrevCursor: ", PrevCursor, ", ");
+            output = string.Concat(output, "RankCount: ", RankCount, ", ");
             output = string.Concat(output, "Records: [", string.Join(", ", Records), "], ");
             return output;
         }
@@ -3182,7 +3222,7 @@ namespace Nakama
         string AvatarUrl { get; }
 
         /// <summary>
-        /// The UNIX time when the user was created.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the user was created.
         /// </summary>
         string CreateTime { get; }
 
@@ -3252,7 +3292,7 @@ namespace Nakama
         string Timezone { get; }
 
         /// <summary>
-        /// The UNIX time when the user was last updated.
+        /// The UNIX time (for gRPC clients) or ISO string (for REST clients) when the user was last updated.
         /// </summary>
         string UpdateTime { get; }
 
@@ -4009,7 +4049,7 @@ namespace Nakama
         string Value { get; }
 
         /// <summary>
-        /// The version hash of the object to check. Possible values are: ["", "*", "#hash#"].
+        /// The version hash of the object to check. Possible values are: ["", "*", "#hash#"].  if-match and if-none-match
         /// </summary>
         string Version { get; }
     }
@@ -4194,7 +4234,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4221,7 +4261,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4248,7 +4288,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4281,7 +4321,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4324,7 +4364,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4372,7 +4412,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4420,7 +4460,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4468,7 +4508,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4520,7 +4560,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4568,7 +4608,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4616,7 +4656,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4664,7 +4704,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4716,7 +4756,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4755,7 +4795,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4789,7 +4829,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4823,7 +4863,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4857,7 +4897,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4895,7 +4935,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4929,7 +4969,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4963,7 +5003,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -4997,7 +5037,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5031,7 +5071,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5066,7 +5106,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5105,7 +5145,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5139,7 +5179,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5173,7 +5213,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5207,7 +5247,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5241,7 +5281,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5275,7 +5315,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5309,7 +5349,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5343,7 +5383,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5377,7 +5417,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5424,7 +5464,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5457,7 +5497,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5496,7 +5536,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5535,7 +5575,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5573,7 +5613,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5610,7 +5650,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5646,7 +5686,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5684,7 +5724,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5737,7 +5777,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5770,7 +5810,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5806,7 +5846,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5844,7 +5884,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5884,7 +5924,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5922,7 +5962,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5960,7 +6000,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -5993,7 +6033,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6031,7 +6071,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6064,7 +6104,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6102,7 +6142,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6147,7 +6187,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6180,7 +6220,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6215,7 +6255,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6250,7 +6290,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6285,7 +6325,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6320,7 +6360,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6355,7 +6395,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6391,7 +6431,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6425,7 +6465,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6475,7 +6515,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6514,7 +6554,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6568,7 +6608,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6620,7 +6660,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6653,7 +6693,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6688,7 +6728,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6732,7 +6772,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6786,7 +6826,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6830,7 +6870,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6864,7 +6904,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6899,7 +6939,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6934,7 +6974,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -6981,7 +7021,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -7029,7 +7069,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -7081,7 +7121,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -7093,6 +7133,39 @@ namespace Nakama
             byte[] content = null;
             var contents = await HttpAdapter.SendAsync(method, uri, headers, content, Timeout, cancellationToken);
             return contents.FromJson<ApiTournamentList>();
+        }
+
+        /// <summary>
+        /// Delete a tournament record.
+        /// </summary>
+        public async Task DeleteTournamentRecordAsync(
+            string bearerToken,
+            string tournamentId,
+            CancellationToken? cancellationToken)
+        {
+            if (tournamentId == null)
+            {
+                throw new ArgumentException("'tournamentId' is required but was null.");
+            }
+
+            var urlpath = "/v2/tournament/{tournamentId}";
+            urlpath = urlpath.Replace("{tournamentId}", Uri.EscapeDataString(tournamentId));
+
+            var queryParams = "";
+
+            var uri = new UriBuilder(_baseUri)
+            {
+                Path = _baseUri.AbsolutePath + urlpath,
+                Query = queryParams
+            }.Uri;
+
+            var method = "DELETE";
+            var headers = new Dictionary<string, string>();
+            var header = string.Concat("Bearer ", bearerToken);
+            headers.Add("Authorization", header);
+
+            byte[] content = null;
+            await HttpAdapter.SendAsync(method, uri, headers, content, Timeout, cancellationToken);
         }
 
         /// <summary>
@@ -7132,7 +7205,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -7171,7 +7244,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -7212,7 +7285,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -7248,7 +7321,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -7299,7 +7372,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -7342,7 +7415,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;
 
@@ -7388,7 +7461,7 @@ namespace Nakama
 
             var uri = new UriBuilder(_baseUri)
             {
-                Path = urlpath,
+                Path = _baseUri.AbsolutePath + urlpath,
                 Query = queryParams
             }.Uri;